Add way to read in Team name and games from directory.
The new (global) function readTeamsFromDir takes a
directory and reads all of the files looking for files with
"games" in the name and ending in ".d". It then parses
the file name to get the name of the team, creates the team
and adds the games in the file.

At the moment, it reads all of the files in properly (I have many
more files than are on github) but ends with a
"double free or corruption (out)" error.  I believe this has
to do with the lack of a copy constructor and proper
destructor in the Team or TeamGame classes.  I will try
to fix later.