Make variables in classes const.
Went through each class and determined which of the
variables in each class could be made const and then
did so.  I also changed the getters to return const
variables as well.

Realized I hadn't removed the "using namespace std" in
Team.h/cpp, so I did that as well.