Fix bug in readTeamsFromDir and clean up main.cpp.
Fixed a bug in readTeamsFromDir by removing the sprintf
function that set path.  Path was an unnecessary variable.
I was able to simply use filepath, which already existed.

I removed most of the testing that was going on in main.cpp.
Now all that is left is the reading in of the files and
testing to ensure that it worked.