Change TeamGame. Addition of generic getPct function.
Changed poss -> oposs to comform to my new standard.

Changed order of declaration for dposs so that it could
be used to properly initialize other member variables in
the constructor.

Added a function called getPct(string which) to TeamGame
to be able to access all of the Pct variables in the class
just by changing the string argument.  This allows me to
cycle through the variables easily.