Complete functionality of gen_averages.cpp.
gen_averages.cpp now runs exactly like the Ruby version
of the program.  It reads in all of the game info and
outputs the average for a bunch of stats (like points,
field goals made, attempted, and percentage, etc).
It then outputs the ratio of the total average to the
average when at home, on the road, and at a neutral
site (the latter of which should be very close to 1 for all).
It also computes the average point differential at the various
locations (neutral is 0 by definition), as well as the ratio of
winning at the various locations (neutral is 1 by definition).

TODO: add game files for more years, change function for
reading in games to accept multiple years, add ability
to decide which years to read in.