Add constants files.
The results generated from gen_averages.cpp are put
into csv files.  There are four files containing the constants.
The format of the files is as follow:

team_5yr_averages: each line contains a single year's worth
of averages.  The line starts with the year the averages
are valid for, followed by the average of points, then
personal fouls (all averages are per-game averages),
then the makes, attempts, and percentage averages for
field goals, two-pointers, three-pointers, free throws,
offensive rebounds, defensive rebounds, total rebounds,
assists, steals, blocks,  turnovers, effective field-goals,
and free throw made rate. The final number is the year
that the line is valid for again.

team_neutral_averages: each line contains a single year's
worth of neutral ratios (total average / home (or away)
average) for home, away, or neutral sites.  Each line
starts with the year the stats are valid for, then the location,
home, away or neutral.  Then the neutral ratios are
presented, in the same order as in the averages file.
Again, the last number in each line is the year it is valid for.

team_point_differentials.d: each line contains a single year's
worth of point differentials at home, away, and at neutral
sites.  The first number is the year, the second is the
point differential at home, the second is the point differential
on the road, and the third is the point differential at
neutral sites, the last number is the year again.  All
of the differentials are averaged over the total
number of games played.  The home/road values are
not equal but opposite because there are a number of
games played against non-Division-I schools where only
one side (usually the home team) is represented.

team_win_ratios.d: each line contains a single year's
worth of win ratios, where the win ratio is average
winning pct (0.5) divided by the winning ratio at home,
away, and at neutral sites.  The first number is the year,
the second is the win ratio at home, the third and fourth
are away and at neutral sites respectively, and the last
number is the year again.

All of the constants were generated from gen_averages.cpp
using the previous 5 year's worth of game data.  Thus,
2006's constants are generated with data from the
2001-2005 seasons.