Change in gen_averages leading to change in constants files.
I required all of the constants to be derived from games
between Division-I schools, so games where one of the
opponents was not a Div-I school (defined to be a school
whose full stats for a particular season I have, which may
slightly differ from the actual values) is now excluded.

This change led to changes in the values of the constants
and thus all of the constants files have been updated.