Add singleton class to contain Team 5 Year Averages.
I am adding in a new singleton class that contains a
hash with two sets of keys, the first being the year
and the second being the type of stat (fgm, fga, fgp, etc).

Because it is a singleton class, it exists when the program
starts, but needs to be initialized to read in the data
that fills the hash.  It is initialized with the command:
ConstantTeam5YearAverages::Instance()->initialize(pathToFile)

The contents can be extracted using the form:
ConstantTeam5YearAverages::Instance()->get(year,stat),
where year is an int in the valid range, and stat is a string
comprising a valid statistic to access (fgm, fga, fgp, etc).