Add singleton class to contain Team Neutral Ratios
I am adding in a new singleton class that contains a
hash with three sets of keys, the first being the year
and the second being the location (home, away, neutral),
and the third being the type of stat (fgm, fga, fgp, etc).

Because it is a singleton class, it exists when the program
starts, but needs to be initialized to read in the data
that fills the hash.  It is initialized with the command:
ConstantTeamNeutralRatios::Instance()->initialize(pathToFile)

The contents can be extracted using the form:
ConstantTeamNeutralRatios::Instance()->get(year,loc,stat),
where year is an int in the valid range, loc is a string
containing home, away, or neutral, and stat is a string
comprising a valid statistic to access (fgm, fga, fgp, etc).