Add singleton class to contain Team Win Ratios
I am adding in a new singleton class that contains a
hash with two sets of keys, the first being the year
and the second being the location (home, away, neutral).

Because it is a singleton class, it exists when the program
starts, but needs to be initialized to read in the data
that fills the hash.  It is initialized with the command:
ConstantTeamWinRatios::Instance()->initialize(pathToFile)

The contents can be extracted using the form:
ConstantTeamWinRatios::Instance()->get(year,loc),
where year is an int in the valid range, and loc is a string
with home, away, or neutral.