Add function to calculate weighted averages.
This is a pretty complex function that is hard to describe, but
the code is heavily commented.  Basically, two sets of
averages are calculated, the raw average for each stat
on a given day (if a team played 3 games, scoring 80, 90, and
100 points, then their average is 90), and then the weighted
average where we take into account the opponent in
each game and how good they are in each stat (those 100
points against a team that averages giving up 120 is not
as impressive).  Thus, each game's stats are weighted
to give a final average that is higher or lower depending
on their opponents.