Add function to get certain values from TeamGames.
In writing the calcWeightedAverage function for Team, I
needed easy access to values, like the offensive field goal
percentage.  So I wrote a function to take a string in the
form "pctName.(m/a/p)", where pctName is the name of
the base stat I want (like offensive field goals) and
m/a/p indicates if I want the number of makes or attempts,
or if I want the percentage.  There are special provisions
for points and personal fouls.  The dot (.) is important
so I don't have to heavily parse the string.

The names in the ConstantTeam5YearAverages and
ConstantTeamNeutralRatios classes were changed to
work with the new naming convention.