Replace insert with emplace for unordered_maps
All constant files had inserts replaced with emplace for
significantly easier code to read.  Also the names of the
variables in ConstantTeam5YearAverages and
ConstantTeamNeutralRatios were changed to include
a dot (.) between the beginning part of the name and
the final m/a/p.  See previous commit for more details.