Add season_info file.
This file contains the info about the start and end of
the NCAA basketball season and the NCAA tournament.
Each line contains one year's worth of data.  The first
number is the year, the next three numbers are the
year, month, and day of the start of the tournament.
The next three are the date of the end of the tournament.
Then the end of the regular season followed by the start
of the regular season.  The ordering may be stupid, but
it doesn't really matter.