Add singleton class to contain Season Info.
I am adding in a new singleton class that contains a
hash with two sets of keys, the first being the year
and the second being a string for the date you want.  The
options are "season start", "season end", "tournament
start", "tournament end".

Because it is a singleton class, it exists when the program
starts, but needs to be initialized to read in the data
that fills the hash.  It is initialized with the command:
ConstantSeasonInfo::Instance()->initialize(pathToFile)

The contents can be extracted using the form:
ConstantSeasonInfo::Instance()->get(year,which_date),
where year is an int in the valid range, and which_date is
a string containing one of the valid options.