New application file, generateWeigtedAverages.cpp
This application will generate the weighted averages files
for every team in a given year, where the year is specified
by the command line switch -y.  This code call
calcWeightedAverages for every day in the season starting
two weeks after the start of the season (to ensure there
are enough games to make weighting worthwhile) until
the start of the NCAA tournament.

This code takes appx 2 minutes per team on my machine
(which isn't particularly fast, so it's possible to run faster),
and with 350 teams, it can take as long as 11-12 hours
to run.  But it only needs to be run once per season.

It results in two files, an averages file, which contains the
raw averages, and a waverages file, which contains the
weighted averages as well as rpi, srs, and sos.  The
files for 2015 north carolina will be uploaded as well.