Change to Team::findteam and addition of addAverages and addWAverages.
findteam now checks to see if the teams map contains
the team so as not to accidentally initialize an entry.

Added full functionality to the addAverage and addWAverage
methods, following the form of addGame.

Then added addAverages and addWAverages to
be able to add all the averages in a file at once, following
the form of addGames.

Some other small changes.