Pointer cleanup in calcWeightedAverage and using typedef statements.
I cycle through the opp_sums and opp_opp_sums at the end of
calcWeightedAverage to delete the pointers created,
which would have created memory issues.  I also
implemented all of the typedefs that had been declared
but not used.  Added typedef for sumsType.

Removed the const qualifier on GameOnDate.