Add new program file to calculate SRS by iteration.
The SRS value is easiest to determine by iteration.  An
initial value is calculated in generateWeightedAverages for
each team on each date during the season.

It is calculated by
srs = average point differential + average opponent srs.
In the initial round the average opponent srs is guessed
at by using their average point differential.  In each round
after the initial calculation though, it is possible to use
the opponents' previously calculated srs values.  After a
few iterations the values should converge to a stable
value.

The program takes two arguments, the year, and the
number of iterations.  Every team in a given year has
its SRS calculated all in one go.  I generally use 20 iterations.

At the moment, it is righting into a new file, but once I've
guaranteed it is working properly, I'll have it just overwrite
the iwaverage file.