Add new program to generate a ROOT file with histograms showing the averages and waverages for a team over the course of the season.
The new program will generate a ROOT file (which requires
ROOT to be installed to be used) with histograms for the
four factors plus SRS and RPI over the course of the season.

The Ruby file is used to generate a nice canvas and actually
display it.  It requires Ruby, libRuby, and ROOT to use.

These files will be updated to show more stats over the coming
days.