Add optional output, option to write to file, and protected against nans in calcSRS.
I added an optional output that takes 3 teams, a generally
good team, a generally mediocre team, and a generally
bad team, and outputs their SRS on four different dates
(14 , 50, and  100 days into the season, and the day
the tournament starts) throughout the season for each
iteration.  This will allow you to determine
the appropriate number of iterations to use.

I added an option to write the output, with the option not
selected, the output will not be written.  This is done to
allow you to check on the number of iterations you feel
is necessary without affecting the actual files since we
are now overwriting the input waverages file.

As mentioned, we now directly overwrite the waverages
files instead of creating a separate file.  This is done
to minimize the number of files (which is already at 3)
associated with each team.

Lastly, I added an extra check for nan values in calcSRS.
This is hopefully unnecessary now that I have implemented
similar changes in calcWeightedAverages in the Team class,
but it makes sense to have the check here as well.