Rename file, change BOOST_FOREACH->for loops, add comments, add output to file.
I renamed the file to be more consistent with the other
programs.  I changed all instances of BOOST_FOREACH to
for loops.  And I added the ability to output the results
directly into files.  You need to specify the name of the
files after the -o switch.  There are four files that need
to be written and so four names need to be provided
in a comma-separated list.