Add histograms for games, add lines with league averages, add comments, remove typedefs.
Added histograms to store game-by-game results, both
weighted and unweighted.  The code for filling them is
provided as well.

Added TLines to hold the lines used to denote the league
averages on the histograms.

Added more comments to make it clearer what is going
on at each step.

Removed the typedefs since they were no longer needed
after changing the BOOST_FOREACH loops to for loops.

Also removed the waverages hash that was not being used.