Create file on fly, plot games, plot lines.
Added ability to plot either from a file or to create the
file on the fly and plot it immediately.  Requires $CLION
environment variable to be set to the location of the binary
for the C++ program.

Added in the ability to plot games now that they're included
in the ROOT file.

Added in the ability to plot the lines for league averages
on the histograms.