Remove typedef, add num_games to averages file, process teams in sorted order.
I removed the typedefs since they are not needed after
switching from BOOST_FOREACH to for loops.

Added num_games to the averages_file for each date.  This
was a bad oversight before.

I sorted the teams before processing so that it is easier
to know how far into the process we are