Change date to non-pointer, remove spread as variable.
Similar change to Game where I changed date from a
pointer to a normal variable.

I decided to remove spread as a variable because it is
easier to just read in the game data if I need the spread.
Also, there are some instances where teams play 2 games
on the same day (generally in pre-season or holiday
tournaments).  It doesn't happen often, but it happens
enough to require this change.  More consequences of
teams playing twice on the same day can be seen in
the next update to Team.