Change to constructor, add variable for origSRS.
Changed constructor to take and pass a normal date
instead of a date pointer due to changes in TeamAverage
seen in last commit.

Also added a variable called origSRS to hold the SRS
calculated by calcWeightedAverage.  SRS is calculated
by iteration.  Once all of the teams have an initial
SRS guess on each date, it is possible to iterate through
and hopefully reach some final value for each team.  But
since that process overwrites the initially calculated value,
I decided to add an extra variable here (and thus to the
files storing the info) so that it could be recalculated
if there were some need to, without having to redo
the entire calcWeightedAverage program again, because
it takes ~11 hours to run for an entire season.