Change GameOnDate, change addGame/Average/WAverage, addGames/Averages/WAverages, change calcWAverage.
The gamesByDate hash is changed to store the games using
a key made up of a date string (of the form xxxx-xx-xx)
and the opponent, instead of just the date.  This was done
because there are a few occurances every year of a team
playing two games in a single day and so one of the games
was being ignored.  This was a problem when iterating
through the SRS values because that iteration depends on
every team and every game being accounted for.

As a result of this change, the GameOnDate method had
to be ammended to take a date and a string, where the
string is the name of the opponent played on that date.
The addGame method also had to be changed to deal
with this.

The addGame/Average/WAverage methods were changed
to reflect date no longer being a pointer.

The addGames/Averages/WAverages methods were changed
to deal with date no longer being a pointer.  The
addAverages/WAverages methods
also were changed to reflect the
fact that spread is no longer a variable.  The
addWAverages method was changed to include
origSRS.

There were also significant changes to
calcWeightedAverage.  Firstly, many comments were
added.  Second, all instances of getDate were changed
to reflect the fact that date is no longer a pointer.
All for loops were changed to have the type be
auto.  The summing of opp_avg_pt_diff was put inside
an if statement requiring num_opp_games[oname] to
be greater than 0.  This will ensure we don't have
nans popping up early in the season.  When
srs and sos are calculated near the end of the
method, they are only inside the num_games > 0 if test,
they had previously been inside another if test, requiring
opp_avg_pt_diff > 0, which was wrong.