Comments added.  Home/away properly set. Recalculation.
The program file itself has not changed much except for
some comments being added.  However, at this point, I fixed
the location of a handful of games which had some
problems.Some of the games had opponent's whose
names were misspelled or inconsistent (milwaukee
instead of wisconsin milwaukee for instance).  Some
 (especially those just fixed) then had inconsistent
 home/away/neutral locations.  These were researched
 to determine the proper location.

After fixing those issues, I was able to determine that in
each year an equal number of home and away games were
played (which is obviously what should be true), as well
as an even number of neutral site games.  The point
differentials are not equal and opposite for home/away,
and neutral is 0.  Likewise the sum of the winnings
percentages at home and away sum to 1.

After these changes, it was necessary to recalculate the
four constants files contents.  These have now been
updated.