Remove writing spread to output files.  Add writing origSRS.
Since Averages/WAverages no longer keep spread as a
variable, it was removed here.  This meant I did not require
getting the TeamGame played on the date.

Since TeamWAverage will now store the initial SRS value,
I write it out twice here.  It seems stupid here, but will
be fine once we iterate, modifying the output file and
leaving us with an initial and final SRS.