Add ability to calculate using origSRS, remove spread from output and add origSRS to output.
Because of the changes with TeamWAverages, spread
is no longer being read in or out, so those lines were
removed.  But I had to add in origSRS to the output.

Also, I add the option to start the calculations from the
original SRS if desired.  This allows us to recalculate
from scratch without having to re-run
generateWeightedAverages, which takes 11+ hours/season.
By default, it will start with the current value.