Change constructors to deal with percentages > 1.
Since I added constructors that allow you to set the
percentage directly, and another one to take the input
makes and attempts as doubles, I needed to ensure that
the percentage never gets above 1.  The weighted averages
at the very beginning of the year can occasionally go too high,
so this safeguards against that.