Add generateWAverageFunctions program.
This program determines a function for how well Team A
should do in each of the "four factors" statistics in a game
against Team B.  Given Team A's offensive weighted
average in the category, and Team B's
defensive weighted average in the same category, we
should be able to predict how well Team A should do.

This program finds those four functions by looking at
every game between two dates (currently the beginning
of December and the beginning of the NCAA tournament)
and tabulating the teams' weighted averages on the
day of the game (thus not including that game), as well
as how Team A actually performed in that game.  So, if
Team A has a weighted average for offensive rebounds of
30%, and Team B has a weighted average of giving up
25% offensive rebounds to their opponent, and Team A
actually got 35% in the game, those numbers are all tallied.

Later, a graph is created with the difference in what Team A
gets and what Team B gives up as the X axis, and what
Team A actually got as the Y axis (actually it's more
complicated than that), and then a linear function is
fit through those points.

The output is a ROOT file with a bunch of histograms and
graphs, as well as the parameters for the fits.
Eventually these will be output to a file, and there will
be a Ruby file to display the graphs and histograms.