Overhaul weighted_std_dev, checks for variance == 0 to several functions.
I added checks to ensure that variance is non-zero to
weighted_average since variance is in the denominator.

I redid weighted_std_dev to conform to a different method.
The main actual change is to divide the sum not by
sum_weights, but by sum_weights times the number
of non-zero weights minus one over the number of
non-zero weights.  This effect should be quite small when
n there are many weights, but is useful for small numbers
of weights.

I also added a check to ensure the variance is non-zero.
And I took the division by sum_weights (and that extra
term) out of the loop, placing it afterwards, to reduce
the number of computations needed.