Significant overhaul.
Removed outYear since all output will be done with
the Ruby file now.

Stopped computing hist_slices because they weren't being
used for anything.  The code is left in place but
commented in case I want to use them later.

Added titles to the histograms and graphs.

Added averagedGraphs, which uses the weighted average
of all the entries at each X value, along with the
error associated with that average (following method
of Taylor).  This required a lot of intermediate
variables, and a new function called
weightedAverageAndError.  These graphs are now
the ones that fit instead of the previous
graphs.

The number of Pcts required in a bin was reduced to
5 from 20.  The previous number was simply a holding
value.  2 would be the bare minimum, but 5 seems
reasonable.

Added comments.