Add Ruby program to plot results of generateWAverageFunctions and deal with output.
Similar to showTeamStats.rb, this program allows you to
look at a file that has already been created or to
create the file on the fly and either save it or discard
it.

Creates single output canvas showing the histograms
created in generateWAverageFunctions as well as the
graphs and respective fits.

It will also output to a text file if desired.

There is code for the hist_slices that is commented
out in case it is desired in the future.