Flip x-axis, add comments, fix incorrect addition.
I decided to flip the x-axis so that the functions can be
evaluated with team A's offense minus team B's defense.
This makes all the slopes negative, but makes the function
evaluations make more sense.  Think of it as defense
is increasing with respect to offense as we move to the
right along the x-axis.

Also, I was adding (range[s]->at(2) - 1) but subtracting
back range[s]->at(2), I corrected that by including the
minus 1.