Add file containing standard deviations of waverages.
These values come from the generateStandardDeviations
program.  Each line contains a year, the name of a stat,
and then 120 values for the standard deviation of that
stat through the 120 days leading up to the start of the
NCAA tournament.  This may seem an odd way to do things,
but it makes the most sense because starting count from
the beginning of the season doesn't work since seasons
may start at a different point relative to the beginning of
the tournament, due to a special game in Japan or something.

A singleton class to hold these values will be added next.