Add singleton class to hold the waverage prediction functions.
These are the values obtained from
generateWAverageFunctions. They are stored in a file
which is used to initialize the class.  This class uses
ROOT TF1 functions, and therefore requires ROOT to
compile and run.

The standard get function in these singleton classes is
replaced by a predict function which takes two waverages,
w1 and w2, and uses their values to predict w1's offensive
output in each of the stats given w1's offense so far and
w2's defense so far.