Add predict function assuming average opponent.
When considering how to rank all teams, it is convenient
to have them play against a fake opponent whose stats
are all the league averages.  This function allows you to
calculate both how w1 would do offensively and how
an average team would do offensively against w1.
The last parameter is optional, it defaults to offense.
The options are "offense" or "defense".