Significant update: complete change to calculations; addition of SRS.
I now calculate the standard deviation for offensive and
defensive versions of the stats and then combine them
in quadrature.  Since I care about these std devs only
for how they relate to the quantity waverage(offensive
stat) - oppWaverage(defensive stat), this is the
most expedient way to do it.

I also added the standard deviation of SRS to the mix
because I need that quantity as well.