Add program generateWeights.
This program calculates the weights to use for each of the
four factors and SRS to obtain the best winning percentage
for my formula, using five (or less) years of data.  These
weights will be used to predict games in the year immediately
following the years used in the study.

Note: there are several ways to run this fit and so there may
be several similar programs available with slightly different
methods.