Complete overhaul.
This program no longer calculates day-by-day, but
instead comes up with a single number per stat per year.
It finds the standard deviation of the difference in the
predicted values of the weighted averages.  I would have liked
to have been able to calculate it from first principles, but
this brute-force method is more accurate.

The change here requred changes in the
ConstantStandardDeviations class, which will be updated next.