Change reading in of values due to change in standard deviations. Add ability to use preset list of starting parameters.
The reading of the values in the global vectors is changed
to handle the new standard deviation method.  Also removed
a few lines and variables needed for the old method, but
not longer used.

I also added the ability to seed the fitter with certain values
instead of using a random number generator.  This
can be used to check the other years' weights to ensure
the best weights are being used, or it could be used
to set up a pre-selected group of values to check so we
don't have to rely on the random number generator
coming up with eough options.  This is useful because we
can pretty well say what the ordering of the importance
of the weights should be, and their relative values,
so a small list could be generated.