Add new program to plot the waverages, srs, rpi, and gameScore for two teams over the course of the season.
Creates a ROOT file with histograms of the waverages, srs,
rpi, and gameScore for each day during the season for
two input teams, and then the results of a mock matchup
between the two assuming the stats of the day of the
beginning of the tournament.

The Ruby program file showTournamentMatchiup.rb is used
to view the results.