Add new class for dealing with regions in the NCAA tournament.
This class is input with all of the teams (at least 16, but
sometimes more if there are play-in games), and then
provides methods for playing each round of the tournament,
starting from the play-in games (if any) and ending up with
a single region champion.  This class should really only
exist in the context of the Tournament class, which will
fill in all of the members and call the methods.  It should not
be called by the user.