Add the tournament information files for each year.
Each file contains 64+ lines, detailing the teams in the
tournament.  Each line contains a region number (from
0 to 3), the name of a team in that region, and the
seed number for that team.  Note: some seeds are
duplicated in a region due to play-in games.

These files are used by tournamentSimulation.cpp and
read in by Tournament using the readTournamentInfo
method.