Significant overhaul.  Add in points per possession, change ranges and how they are used.
I had been using the entire range from 0 to 1 for each of the
four factors, which made some of the calculations a bit
simplified, since I had to subtract off 0 and divide by 1.
I changed the ranges to more accurately reflect where the
values may fall, and so had to change the functions
which determine which bin an event will fall into.

Also added points per possession and treated it a bit
differently because its values can be above 1, which is
problematic when treating them like normal Pct values.