Add ability to append to files and select date range.
Instead of just overwriting or skipping a file, you may now
choose to append to a file to add dates later in the season
without recalculating all dates from scratch.  This is useful
for dealing with the current season.

You may also choose a date range using the command line
as well.  The default is to start two weeks after the start
of the season and to end at the beginning of the tournament.