Change way dates are processed.
Previously, I iterated over the dates from 14 days after the
start of the season until the start of the tournament,
now I look at a generic team (North Carolina in any given
season) and process all of the dates that it has waverages
for.  This handles the fact that I generate averages and
waverages for the last day of the tournament separately
from the days before the tournament.