Renamed from showTournamentMatchup.cpp. Significant changes.
Add ability to select date of matchup so it doesn't have to
be only on the first day of the tournament.  Add ability
to run without ROOT output, giving only the text output,
which was just added to say which team is more likely
to win the game.  Add separate histograms to hold the
predicted values so that I can make them look different
when I plot them.