Add class to hold scheduled games for a team.
This class holds very basic information about games to be
played over the course of the season for a team.  The fields
are the name of the team and its opponent, the game
number in the season, the date of the game, and the
location of the game.  Only games against division-I
opponents are included.  Only scheduled games are
included, thus if there is a tournament (early-season
tournaments are common) then only the first-scheduled
game will be listed, since you can't know who they'll play
in the second round or beyond.