Add has of TeamScheduledGames and getters/setters.
With the addition of the TeamScheduledGame class, it is
necessary to have a container to hold those games in
the Team class.  They are stored in a hash indexed by the
date (in string form) with values of a vector of
TeamScheduledGames; the vector is used because it is
possible to have two games scheduled on the same day,
though it is rare.

Methods are created for adding in a scheduled game,
adding in an entire file of scheduled games, accessing the
hash of scheduled games, and accessing the vector
of scheduled games on a given date.