Add helper functions to convert between predicted winning percentages, spreads, and moneylines.
The conversion from predicted winning percentage to
moneyline is known formula.  The conversion from
moneyline to spread is not exact, but generally follows
an exponential function (assuming the moneylines are
kept all as positive values instead of using negative
values for favorites), which was determined by plotting a
bunch of moneylines vs their corresponding spread values.
The inverse of that function can be determined from
basic arithmetic.  Conversions between spreads and
winning percentages use the moneyline as an intermediary.