Add program to predict the winners of all games scheduled to be played on a date.
Unlike checkGamePredictions, which analyzes games that
have been played already, this program looks at every
team in memory and if they have a game scheduled for
the set date, makes a prediction for the game based on
the waverages on that date (or another date selected).
The output is a line saying which of the two teams is favored
to win the game, and then a line for each team in the game
stating the moneyline and spread that should be assigned
to each team based on their predicted winning percentage.

In reality, the spreads are generally equal and opposite,
while the moneylines can be opposite, but not equal.
However, in this program, the fact that the winning
percentages must add to 1 and the way the moneylines
are calculated from those percentages ensures that
the moneylines are roughly equal and opposite, while the
spreads are opposite but not equal.