Update to allow locking SRS value to zero if desired.
This basically eliminates the need for the FixedSRS version
of generateWeights.  While this can only lock the SRS value
to 0, that is the only value it is really necessary to lock it to.
Putting it here is both tidier and considerably faster since
it is not necessary to calculate the norm at every step.

I will likely leave generateWeightsFixedSRS around in case
I want to lock the SRS to other values, but 0 is the most
common number to lock the SRS to.