Update all functions to handle favorites properly.
Vegas works in an odd way.  The underdog's likely winning
percentage is directly related to their spread and their
moneyline, but the favorite's spread is defined to be the
opposite of the underdog's, which does not lead to a
direct conversion from winning percentage to moneyline
or spread.  These changes handle that.