Change calculateTotalPercentages function.
There were three big mistakes.  The first two were that
teamB in both the Final Four and Championship game
matchups were accidentally being set to the same team
as teamA, which is a pretty big problem.  The other was
that the opposite regions were messed up for regions 2 and 3
which were set to play themselves instead of regions 0 and 1.
All has been fixed.